+++ .github/workflows/coverage.yml
@@ -13,13 +13,18 @@ jobs:
       - name: Install
         run: npm install
       - name: Test and Coverage
-        run: npm run 'test:cov'  # or npm run coverage
+        run: npm run 'test:cov'
 
-      # Add this
-      - name: Update Coverage Badge
-        # GitHub actions: default branch variable
-        # https://stackoverflow.com/questions/64781462/github-actions-default-branch-variable
-        if: github.ref == format('refs/heads/{0}', github.event.repository.default_branch)
-        uses: we-cli/coverage-badge-action@main
+      - name: Create Coverage Badges
+        uses: jaywcjlove/coverage-badges-cli@main
         with:
-          node-version: 24.x
+          style: flat
+          source: coverage/coverage-summary.json
+          output: coverage/badges.svg
+          jsonPath: totals.percent_covered
+
+      - name: Deploy
+        uses: peaceiris/actions-gh-pages@v4
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          publish_dir: ./badges
