+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# v1.4.10
+
+- failure to convert colors to display-py-linear du to missing import
+
 # v1.4.9
 
 ## Improvements
