+++ jsr.json
@@ -1,6 +1,6 @@
 {
   "name": "@tbela99/css-parser",
-  "version": "1.4.9",
+  "version": "1.4.10",
   "publish": {
     "include": [
       "src",
