+++ src/lib/syntax/color/utils/distance.ts
@@ -1,8 +1,9 @@
 import type { ColorToken } from "../../../../@types/index.d.ts";
-import { convertColor, toPrecisionValue } from "../color.ts";
+import { convertColor } from "../color.ts";
 import { getOKLABComponents } from "../oklab.ts";
 import { ColorType } from "../../../ast/types.ts";
 import { colorDistancePrecision } from "../../constants.ts";
+import { toPrecisionValue } from "../../syntax.ts";
 
 /**
  * Calculate the distance between two okLab colors.
@@ -31,7 +32,6 @@ export function okLabDistance(color1: ColorToken, color2: ColorToken): number | 
 
     // include alpha
     if (okLab1[3] != null || okLab2[3] != null) {
-        
         diff.push((okLab1[3] ?? 1) - (okLab2[3] ?? 1));
     }
