+++ src/lib/syntax/syntax.ts
@@ -33,6 +33,9 @@ import {
     COLORS_NAMES,
     colorFuncColorSpace,
     LOC,
+    anglePrecision,
+    colorPrecision,
+    epsilon,
 } from "./constants.ts";
 import { getSyntaxConfig } from "../validation/config.ts";
 
@@ -1847,4 +1850,108 @@ export function isValue(token: Token) {
         token.typ === EnumToken.MathFunctionTokenType ||
         token.typ === EnumToken.TransformFunctionTokenType
     );
+}
+
+// https://developer.mozilla.org/en-US/docs/Learn_web_development/Core/Styling_basics/Values_and_units#absolute_length_units
+/**
+ * Convert length to px
+ * @param value
+ * @returns
+ */
+export function length2Px(value: LengthToken | NumberToken): number | null {
+    let result: number | null = null;
+    if (value.typ == EnumToken.NumberTokenType) {
+        result = +value.val;
+    } else {
+        switch ((value as LengthToken).unit) {
+            case "cm":
+                // @ts-ignore
+                result = (value as LengthToken).val * 37.8;
+                break;
+            case "mm":
+                // @ts-ignore
+                result = (value as LengthToken).val * 3.78;
+                break;
+            case "Q":
+                // @ts-ignore
+                result = ((value as LengthToken).val * 37.8) / 40;
+                break;
+            case "in":
+                // @ts-ignore
+                result = (value as LengthToken).val / 96;
+                break;
+            case "pc":
+                // @ts-ignore
+                result = (value as LengthToken).val / 16;
+                break;
+            case "pt":
+                // @ts-ignore
+                result = ((value as LengthToken).val * 4) / 3;
+                break;
+            case "px":
+                result = +(value as LengthToken).val;
+                break;
+        }
+    }
+
+    return isNaN(result as number) ? null : result;
+}
+
+/**
+ * minify number
+ * @param val
+ */
+
+export function minifyNumber(val: string | number): string {
+    val = String(toPrecisionValue(val));
+
+    if (val === "0") {
+        return "0";
+    }
+
+    const chr: string = val.charAt(0);
+
+    if (chr == "-") {
+        const slice: string = val.slice(0, 2);
+
+        if (slice == "-0") {
+            return val.length == 2 ? "0" : "-" + val.slice(2);
+        }
+    }
+
+    if (chr == "0") {
+        return val.slice(1);
+    }
+
+    return val;
+}
+
+export function toPrecisionValue(value: number | string, precision: number = colorPrecision): number {
+    const div: number = Math.pow(10, precision);
+    // @ts-ignore
+    value = Math.round(value * div) / div;
+
+    return Math.abs(value) < epsilon ? 0 : value;
+}
+
+export function toPrecisionAngle(
+    angle: number,
+    precision: number = colorPrecision,
+    correctValue: boolean = true,
+): number {
+    angle = toPrecisionValue(angle, precision);
+
+    if (correctValue && Math.abs(angle) >= 360) {
+        angle %= 360;
+    }
+
+    if (Math.abs(angle) < anglePrecision) {
+        angle = 0;
+    }
+
+    if (correctValue && angle < 0) {
+        angle += 360;
+    }
+
+    return angle;
 }
